# perf: fix the performance issue of `append_paged_kv_cache`