--- a/include/flashinfer/page.cuh
+++ b/include/flashinfer/page.cuh
@@ -271,9 +271,8 @@ __global__ void AppendPagedKVCacheKernel(paged_kv_t<DType, IdType> paged_kv,
     uint32_t page_iter, entry_idx;
     paged_kv.page_size.divmod(paged_kv.indptr[batch_indices[i]] * paged_kv.page_size + positions[i],
                               page_iter, entry_idx);
-    size_t kv_offset = paged_kv.get_elem_offset(page_iter, head_idx, entry_idx, tx * vec_size);
-    DType* k_ptr = paged_kv.k_data + kv_offset;
-    DType* v_ptr = paged_kv.v_data + kv_offset;
+    DType* k_ptr = paged_kv.get_k_ptr(page_iter, head_idx, entry_idx, tx * vec_size);
+    DType* v_ptr = paged_kv.get_v_ptr(page_iter, head_idx, entry_idx, tx * vec_size);
     vec_t<DType, vec_size>::memcpy(
         k_ptr, append_key + i * append_k_stride_n + head_idx * append_k_stride_h + tx * vec_size);
     vec_t<DType, vec_size>::memcpy(