--- a/tests/test_page.py
+++ b/tests/test_page.py
@@ -14,6 +14,7 @@ def test_append_paged_kv_cache():
     kv_append_indptr = torch.cat(
         [torch.zeros(1).int().to(0), torch.cumsum(kv_append_length, dim=0)]
     ).int()
+
     max_num_pages = 1000
     page_size = 16
     paged_kv_cache = (
@@ -30,11 +31,17 @@ def test_append_paged_kv_cache():
     # 25 = (2 - 1) * 16 + 9
     # 22 = (2 - 1) * 16 + 6
     kv_last_page_len = torch.tensor([13, 8, 9, 6], dtype=torch.int32, device="cuda:0")
+    batch_indices, positions = flashinfer.page.get_batch_indices_positions(
+        kv_append_indptr,
+        flashinfer.page.get_seq_lens(kv_append_indptr, kv_last_page_len, page_size),
+        nnz_kv,
+    )
 
     flashinfer.page.append_paged_kv_cache(
         k_append,
         v_append,
-        kv_append_indptr,
+        batch_indices,
+        positions,
         paged_kv_cache,
         kv_page_indices,
         kv_page_indptr,